fix: Set value into correct field for field transformers (#3646)
This PR fixes an issue in the field transformers where certain fields
were not being correctly assigned when set ddl options.
The bug specifically affected cases where both `equals` and
`parentheses` were set on a field.

The only affected case found was in `secrets_def.go`, so the DDL was
regenerated for that file using the SDK.


## Test Plan
* [x] integration tests for `secrets_gen.go`