fix: Fix parsing username in handling grants (#3688)
Recently, we introduced a fix in handling users' grants, which addressed
changes in the `2025_02` bundle. The username was parsed incorrectly if
it had a prefix formed of `U`, `S`, `E`, and `R` characters. The
username returned from `SHOW GRANTS` was incorrect in this case. Now,
such names should be handled correctly.

Additionally, fix the parameter order in `assert.Equal` in a related
test.