feat: get datasource tables on par with views (#3249)
When using the terraform plugin we've got hindered by the discrepancy
between the datasources views and tables. It seems the datasource views
has already been further developed than tables. In the roadmap it is
stated that the redesign for snowflake_tables is in progress, but for us
this is a blocking issue, hence this MR.
The goal of this MR is to add the new functionalities of the views
datasource to tables as well.

## Test Plan
* [ ] acceptance tests for changed datasource Tables have been changed
* [ ] integration tests for tables have been changed as filter types
have been added and changed

## References
*
https://registry.terraform.io/providers/Snowflake-Labs/snowflake/latest/docs/data-sources/views
*
https://registry.terraform.io/providers/Snowflake-Labs/snowflake/latest/docs/data-sources/tables
*
https://github.com/Snowflake-Labs/terraform-provider-snowflake/blob/main/v1-preparations/LIST_OF_STABLE_RESOURCES_FOR_V1.md

---------

Co-authored-by: = <hpors@anwb.nl>