docs: remove false attribute from snowflake_shared_database example (#3744)
The snowflake_shared_database example includes the "is_transient"
attribute in the resource definition even though this resource does not
expose that attribute. This PR updates the example & doc to fix that.

Update: The previous PR
(https://github.com/snowflakedb/terraform-provider-snowflake/pull/3743)
was based on the main branch rather than dev; this PR uses the dev
branch for the PR instead.

## References

*
https://github.com/snowflakedb/terraform-provider-snowflake/issues/3731