fix: Fix data types parsing for functions and procedures with 2025_03 Bundle (#3827)
This PR addresses two BCRs in the 2025_03 Bundle:
-
https://docs.snowflake.com/en/release-notes/bcr-bundles/2025_03/bcr-1942
-
https://docs.snowflake.com/en/release-notes/bcr-bundles/2025_03/bcr-1944

The main problem was the changed argument signature from the SHOW
command output. The old parsing function could not handle the
parentheses, causing the mapping to fail. It caused:
-
[snowflake_functions](https://registry.terraform.io/providers/snowflakedb/snowflake/latest/docs/data-sources/functions)
and
[snowflake_procedures](https://registry.terraform.io/providers/snowflakedb/snowflake/latest/docs/data-sources/procedures)
to be inoperable
- all function and all procedure resources to fail reading their state
from Snowflake, which led to removing them from terraform state (if
`terraform apply` or `terraform plan --refresh-only` were run.

The other problem did not manifest in resources or data sources,
however, we need to adjust our docs (and potentially future logic) in
context of data types defaults hardcoded in the provider.

- Remove incorrect `ParseFunctionArgumentsFromString` function
- Use `ParseFunctionAndProcedureArguments` instead to parse function
arguments
- Update expectations in function, procedure, and DMF integration tests,
caused by 2025_03 Bundle activation
- Verify function and procedure behavior when we don't fill the data
type sent to Snowflake with the prebaked defaults
- Add test cases for new max sizes
- Verify new defaults for `VARCHAR` and `BINARY` types for functions and
procedures
- Add acceptance tests proving the correctness of the adjustments
- Add `ReturnTypeOld` for more straightforward logic in data sources
- Update show output schemas for function and procedure
- Generate model builders for function and procedure data sources
- `LegacyDataTypeWithAttrs` and `LegacyDataTypeWithAttrsCanonical` were
added to be able to verify the new data type return format from
Snowflake
- Add new max values for `VARCHAR` and `BINARY` types
- Fix git repositories acceptance test expectations
- Add function and procedure creation in test clients

Next PR:
- migration guide entry and documentation clarifications (new defaults
in Snowflake for VARCHAR and BINARY, and the current approach)
- (optional) Describe manual reproduction steps

Later work:
- Change `convert()` signature inside SDK to return mapping errors
- Change `NewSchemaObjectIdentifierWithArguments` signature inside the
SDK to return an error if data types cannot be parsed

References:
https://github.com/snowflakedb/terraform-provider-snowflake/issues/3822
https://github.com/snowflakedb/terraform-provider-snowflake/issues/3823