chore: Add acceptance tests for rotating pats with an external resource (#3871)
Add acceptance tests for rotating pats with an external resource.

We already provided examples in the documentation using the
`time_rotating` resource, but we only had acceptance tests with manual
triggers. This change adds a few steps showing the behavior with the
`time_rotating` resource and also tests when `keepers` are known after
apply.

## References

https://registry.terraform.io/providers/hashicorp/time/latest/docs/resources/rotating