chore: Listing basic resource with autogenerated content (#3885)
## Changes
- Prepare a schema only with name type, and basic validation
- Generate model, assertions, and show schema
- Important note about model: It was generated with dry-run and copied
manually into a file that doesn't end with _gen (because otherwise it
would be removed when calling generate). It was done because the
generator was throwing errors, because the listing schema contains a
list that is required, which produced invalid model constructors.
- This can also be a good place to discuss if we would like to go with
this schema, my take on the current one:
- There are two/three main things to take into consideration (manifest
and its versioning, listing state)
- Manifest can be either sourced from a stage or inlined as a string.
The source can be freely altered (e.g. going from inlined to stage and
vice-versa). To add a new version in inlined fashion, you can just
change the value in the configuration, and we would call ALTER LISTING
AS and pass the new manifest (the Snowflake then generated a new version
automatically, but without version alias). When the manifest is sourced
from the stage a new version is explicitly created with version name (in
`SHOW VERSIONS IN LISTING` it's returned as `alias`). A new version will
be triggered by:
       - Specifying a new version name
       - Going from inlined approach to the stage one
       - Whenever the stage identifier or stage location changes
- The inlined and stage versions differ in SQL (I'm talking about ALTER
AS vs ALTER ADD VERSION). The inlined one can directly specify if the
new version should be reviewed/published, but the stage version does not
have it. I already asked the listing team how this functionality
corresponds to the stage version. I'm guessing that calling separate
ALTERs with REVIEW and/or PUBLISH commands should be similar to `ALTER
LISTING AS .. REVIEW = x PUBLISH = y' behavior. If my assumptions are
correct, then the top-level review and publish "booleans" should be
sufficient as they will be referring to the state of the latest version.
We could also specify this as a limitation, that the listing state
(review and publish state) will be only considered for the latest
version. Seeing our customers' usage we could extend its functionality.

## Next prs
- Update the schema (add/update descriptions / diff suppresses /
validations)
- Fill the resource implementation
- Add acceptance tests

Reference: [ALTER
LISTING](https://docs.snowflake.com/en/sql-reference/sql/alter-listing)