fix: Change convert signature for a single object (#3926)
It's the first PR of a few aiming to improve SDK's stability and
reliability by including an error in the return from conversion attempt.
Till now, the errors there were only logged which caused unexpected
behavior in some cases (like improper id parsing for functions and ,
because of that, not finding the function with an empty id).

The change contains:
- Rename `convertibleRow` to `convertibleRowDeprecated`
- Introduce `conversionErrorWrapped` method (it's needed as `convert` is
either used from `convertRows` or directly)
- Alter the conversion for warehouse
- Add unit tests for warehouse conversion

Next PRs:
- Migrate all other objects that are not generated from defs
- Alter generation
- Migrate all other objects that are generated
- Rename `convertRowsErr` to `convertRows` when all objects are migrated
- Rename `convertErr` to `convert` when all objects are migrated
- Remove `convertibleRowDeprecated` when all objects are migrated

Future improvements:
- Add compile checks like `var _ convertibleRowDeprecated[DatabaseRole]
= new(databaseRoleDBRow)` for each convertible (generate it)

Questions:
- Do we want to have extensive unit tests for each conversion?
- Do we want to generate these tests?